Add files via upload
*Description: Conducting an an A/B test to compare the effectiveness of gain vs loss framing
messages in registrations into an agricultural information system uisng administrative data with basic sociodemographic characteristics,
and data collected from the registration survey for the SMS campaign. The registration survey data has
information on the type of invitation SMS, respondents’ reply to the invitation SMS and response to the SMS
about the crops they would grow that season.