Add files via upload
*Description: This is an evaluation for an intervantion that intends to increase the uptake of  XYZ services. 
This is not an experimental study and rather an observational one where in a pre and post design is levraged to assess 
the impact based on two rounds of data collection (Baseline and Midline).
Due to the design of this approach, the sample repsondents at basline will be compared to sample repsondents at 
midline (i.e. respondents at basline will act as a counterfactual) along key demographic such as age, income, education, marital status etc) 
and contributing variables that are likely to influence the outcome variables.